# Conflicts:
#	app/build.gradle

+++ app/src/main/java/models/CompassDirection.java
@@ -4,12 +4,12 @@ public class CompassDirection {
 
     float direction;
 
-    public CompassDirection(int direction){
+    public CompassDirection(float direction){
 
         setDirection(direction);
     }
 
-    public void setDirection(int direction) {
+    public void setDirection(float direction) {
 
         this.direction = direction;
     }
